refactor: hook that was un utils moved to hooks folder
