+++ src/hooks/useSafeState.js
